Update README.md
Novel coronavirus pandemic hit the globe in December 2019. It caused a health 
emergency globally. The virus has caused millions of deaths all over the globe. This 
has caused tremendous pressure on the healthcare system and social lives. Thus, the 
need to detect COVID at an early stage and also at a faster rate will help us fight against 
the virus. One of the primary reasons for the rapid spreading of the disease is not having 
the testing kits and also the time it takes to give the result of the test. Hence, using the 
imaging techniques like chest computed tomography scans or chest X-rays will detect 
the virus quickly and effectively as lungs are affected when a person comes in contact 
with the virus. In our work, we used chest X-rays to detect COVID-19 rather than chest 
computed tomography scans as chest X-rays are more affordable and available in every 
clinic. We have used the convolutional neural network (CNN) algorithm to predict the
virus. We built four different models i.e., Xception, VGG19, Inception V3 and 
ResNet50. We determined the results, confusion matrices, and accuracy of all models. 
Xception model gave out the best accuracy with a 93%.